Renamed several services
