+++ src/background/background.ts
@@ -1,6 +1,6 @@
-import "./services/camera_bubble_controller/camera_bubble_controller";
-import "./services/current_tab_setter/current_tab_setter";
-import "./services/current_window_setter/current_window_setter";
-import "./services/permissions_checker/permissions_checker";
+import "./services/active_tab_setter/active_tab_setter";
+import "./services/active_window_setter/active_window_setter";
+import "./services/camera_bubble_injector/camera_bubble_injector";
+import "./services/permissions_giver/permissions_giver";
 import "./services/recording_controller/recording_controller";
 import "./services/storage_controller/storage_controller";
