+++ src/background/services/active_window_setter/active_window_setter.ts
