+++ src/background/services/camera_bubble_injector/injector.ts
