+++ src/background/services/permissions_giver/permissions_giver.ts
